requirements.txt
contourpy==1.3.0
cycler==0.12.1
einops==0.8.1
filelock==3.18.0
fonttools==4.56.0
fsspec==2025.3.1
imageio==2.37.0
importlib_resources==6.5.2
Jinja2==3.1.6
kiwisolver==1.4.7
lazy_loader==0.4
MarkupSafe==3.0.2
matplotlib==3.9.4
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
packaging==24.2
pillow==11.1.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
scikit-image==0.24.0
scipy==1.13.1
six==1.17.0
sympy==1.13.1
tifffile==2024.8.30
torch==2.6.0
torchvision==0.21.0
typing_extensions==4.13.0
zipp==3.21.0
